+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `@metamask/gas-fee-controller` (^26.0.0)
     - `@metamask/network-controller` (^26.0.0)
     - `@metamask/remote-feature-flag-controller` (^2.0.1)
-    - `@metamask/transaction-controller` (^62.0.0)
+    - `@metamask/transaction-controller` (^62.1.0)
   - In clients, it is now possible for multiple versions of these packages to exist in the dependency tree.
     - For example, this scenario would be valid: a client relies on `@metamask/controller-a` 1.0.0 and `@metamask/controller-b` 1.0.0, and `@metamask/controller-b` depends on `@metamask/controller-a` 1.1.0.
   - Note, however, that the versions specified in the client's `package.json` always "win", and you are expected to keep them up to date so as not to break controller and service intercommunication.